# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduConnect API - connect students with tutors
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/threads/{tutorId}:
    get:
      operationId: getMessageThread
      tags: [messages]
      parameters:
        - name: tutorId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Messages in thread with this tutor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      parameters:
        - name: tutorId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessagePayload"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /marketplace:
    get:
      operationId: listMarketplaceItems
      tags: [marketplace]
      summary: List marketplace items with optional filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [popular, newest, price-asc, free-first] }
      responses:
        "200":
          description: List of marketplace items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
  /marketplace/{id}:
    get:
      operationId: getMarketplaceItem
      tags: [marketplace]
      summary: Get a single marketplace item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
...
  /marketplace/{id}/purchase:
    post:
      operationId: purchaseMarketplaceItem
      tags: [marketplace]
      summary: Purchase an item using credits
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplacePurchaseResult"
        "400":
          description: Insufficient credits or error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
components:
  schemas:
...
    Message:
      type: object
      properties:
        id: { type: string }
        tutorId: { type: string }
        sender:
          type: string
          enum: [student, tutor]
        content: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, tutorId, sender, content, createdAt]
    SendMessagePayload:
      type: object
      properties:
        content: { type: string, minLength: 1 }
      required: [content]
    Notification:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [session_reminder, message, booking_confirmed, review_request, system]
        title: { type: string }
        body: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
        link: { type: string }
      required: [id, type, title, body, read, createdAt, link]
...
    MarketplaceItem:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        creatorName: { type: string }
        creatorAvatar: { type: string, nullable: true }
        category: { type: string }
        priceCredits: { type: integer }
        isFree: { type: boolean }
        rating: { type: number }
        reviewCount: { type: integer }
        downloads: { type: integer }
        description: { type: string }
        previewUrl: { type: string, nullable: true }
        fileType: { type: string }
        isFeatured: { type: boolean }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, title, creatorName, category, priceCredits, isFree, rating, reviewCount, downloads, description, fileType, isFeatured, tags, createdAt]
...
    MarketplacePurchaseResult:
      type: object
      properties:
        success: { type: boolean }
        newBalance: { type: integer }
        message: { type: string }
      required: [success, newBalance, message]
...